Remove dependency on parsy
